requirements.txt
asgiref==3.7.2
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
Django==3.2
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.4.2
firebase-admin==6.2.0
google-api-core==2.14.0
google-api-python-client==2.109.0
google-auth==2.24.0
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.13.1
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpcio==1.59.3
grpcio-status==1.59.3
httplib2==0.22.0
idna==3.6
install==1.3.5
msgpack==1.0.7
proto-plus==1.22.3
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pymongo==3.12.1
pyparsing==3.1.1
pytz==2023.3.post1
requests==2.31.0
rsa==4.9
sqlparse==0.2.4
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.1.0
